moved schema to config folder
